interceptor/timeout/interceptor.go
// Copyright (c) 2021 rookie-ninja
//
// Use of this source code is governed by an Apache-style
// license that can be found in the LICENSE file.

// Package rkgintimeout is a middleware of gin framework for timing out request in RPC response
package rkgintimeout

import (
	"github.com/gin-gonic/gin"
	"github.com/rookie-ninja/rk-gin/interceptor"
)

// Interceptor Add timeout interceptors.
func Interceptor(opts ...Option) gin.HandlerFunc {
	set := newOptionSet(opts...)

	return func(ctx *gin.Context) {
		ctx.Set(rkgininter.RpcEntryNameKey, set.EntryName)

		set.Tick(ctx, ctx.Request.URL.Path)
	}
}
